+++ CHANGELOG.md
@@ -5,6 +5,10 @@ All notable changes to this project will be documented in this file.
 The format is mainly based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## Unreleased
+
+- Fixed: Corrected the container image path to ghcr.io/claudiospizzi/ansiblectl in the Start-AnsibleCtl cmdlet
+
 ## 1.0.0 - 2025-09-19
 
 - Added: Initial development release of AnsibleCtl container image with the Start-AnsibleCtl cmdlet